+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 ## [2.29.2] - 2021-06-11
+### Added
+- GMUHK harvester
 ### Changed
 - Enable Edu section for public (in main nav)